requirements.txt
langchain>=0.0.267
langchain-community>=0.0.10
langchain-core>=0.1.0
pillow>=9.5.0
requests>=2.31.0
transformers>=4.30.2
torch>=2.0.1
torchvision>=0.15.2
fastapi>=0.100.0
uvicorn>=0.23.1
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
accelerate>=0.21.0
sentencepiece>=0.1.99
protobuf>=4.23.4
